refactor channels
