+++ qiskit/pulse/channels/output_channels.py
@@ -8,8 +8,40 @@
 """
 Channels support signal output.
 """
+from abc import abstractmethod
 
-from .backend_channels import OutputChannel
+from qiskit.pulse import commands
+from .pulse_channel import PulseChannel
+
+
+class OutputChannel(PulseChannel):
+    """Output Channel."""
+
+    supported = (commands.FrameChange,
+                 commands.FunctionalPulse,
+                 commands.PersistentValue,
+                 commands.SamplePulse)
+
+    @abstractmethod
+    def __init__(self, index: int, lo_frequency: float = None):
+        super().__init__(index)
+        self.lo_frequency = lo_frequency
+
+    def __eq__(self, other):
+        """Two output channels are the same if they are of the same type, and
+        have the same index and lo_frequency.
+
+        Args:
+            other (OutputChannel): other OutputChannel
+
+        Returns:
+            bool: are self and other equal.
+        """
+        if type(self) is type(other) and \
+                self.index == other.index and \
+                self.lo_frequency == other.lo_frequency:
+            return True
+        return False
 
 
 class DriveChannel(OutputChannel):
